fix bls tests finally
