+++ shared/bls/BUILD.bazel
@@ -2,19 +2,22 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = ["core_functions.go"],
+    srcs = [
+        "bls.go",
+        "mcl.go",
+    ],
     cdeps = [
         "@herumi_mcl//:mcl-lib",
     ],
     cgo = True,
-    clinkopts = ["-lstdc++ -lgmp"],
-    copts = ["-DMCLBN_FP_UNIT_SIZE=6 -DMCL_DONT_USE_OPENSSL"],
+    clinkopts = ["-lbls384_dy -lcrypto -lgmp -lgmpxx -lstdc++"],
+    copts = ["-DMCLBN_FP_UNIT_SIZE=6"],
     importpath = "github.com/prysmaticlabs/prysm/shared/bls",
     visibility = ["//visibility:public"],
 )
 
 go_test(
     name = "go_default_test",
-    srcs = ["core_functions_test.go"],
+    srcs = ["bls_test.go"],
     embed = [":go_default_library"],
 )
